requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
cffi==1.17.1
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
dlib==19.24.6
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==24.3.25
fonttools==4.54.1
flask==2.2.2
flask-cors==2.1.2
flask-restful==0.3.9
flask-login==0.6.2
flask-wtf==1.0.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
ml_dtypes==0.5.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sounddevice==0.5.1
stack-data==0.6.3
tensorflow=2.15.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
widgetsnbextension==4.0.13
